refactor(utils): move utils file to utils directory and rename to utils_parser
- moved the `utils` file to a new `utils` directory
- renamed the file to `utils_parser.py`
- updated all relevant imports to reflect the new path and file name
